requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.10b0
click==8.0.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
jsonschema==4.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylama==7.7.1
pymongo==3.12.1
pyrsistent==0.18.0
pytz==2021.3
regex==2021.11.2
snowballstemmer==2.1.0
sqlparse==0.2.4
tomli==1.2.2
typing-extensions==3.10.0.2
